♻️ refactor: Providerの整理
