+++ src/pages/_app.page.tsx
@@ -3,13 +3,11 @@ import "@/styles/global.css";
 import type { AppProps } from "next/app";
 import { DefaultSeo } from "next-seo";
 import SEO from "next-seo.config";
-import { ThemeProvider } from "next-themes";
 
-import { SearchProvider } from "@/components/organisms/SearchArea/SearchContext";
+import { AppProviders } from "@/contexts";
 import { GoogleAnalytics } from "@/lib/google-analytics";
 import { notoSansJP } from "@/styles/font";
 
-// FIXME: MSW is broken "ERR_UNSUPPORTED_DIR_IMPORT"
 // if (process.env.NEXT_PUBLIC_MSW_ENABLED === "true") {
 //   (async () => {
 //     const { setupMSW } = await import("../../mocks/msw/worker");
@@ -23,13 +21,11 @@ const App = ({ Component, pageProps }: AppProps) => {
     <>
       <DefaultSeo {...SEO} />
       <GoogleAnalytics />
-      <ThemeProvider attribute="class" enableSystem>
-        <SearchProvider>
-          <div className={`${notoSansJP.variable} font-sans`}>
-            <Component {...pageProps} />
-          </div>
-        </SearchProvider>
-      </ThemeProvider>
+      <AppProviders>
+        <div className={`${notoSansJP.variable} font-sans`}>
+          <Component {...pageProps} />
+        </div>
+      </AppProviders>
     </>
   );
 };
